Language specifig module definitions
Add module_commons for common structure
Module class provides language specific flavors based on 'lang'
parameter.

verilog_module and vhdl_entity overload the properties and methods
of module_commons.
